requirements.txt
numpy
pandas
scikit-learn
xgboost
torch
streamlit
requests
plotly
matplotlib
seaborn
python-dotenv
pyyaml
joblib
hopsworks